# metrics: optimize rpc latency related metrics