| Workload | Threads | QPS (this PR) | QPS (baseline) | Diff |
| -------- | ------: | ------------: | -------------: | ---: |
| pointget |      64 |      88416.74 |       87059.32 | 1.6% |
| batchget |      64 |      22450.64 |       21999.43 | 2.1% |

This PR
![2025-01-15_093238](https://github.com/user-attachments/assets/80b409ad-9ccf-40bc-8e6d-5e183c9f91eb)

Baseline
![2025-01-15_093250](https://github.com/user-attachments/assets/7d6b1dd3-eaf2-4ddb-8d75-d0972b96cb0b)